timeStamp changed to double

+++ meetmeserver/src/main/java/de/dhbw/meetme/rest/GPSService.java
@@ -50,7 +50,7 @@ public class GPSService {
             activeUser.setLatitude(lat);
             activeUser.setLongitude(lon);
 
-            long currentMillis = System.currentTimeMillis();
+            double currentMillis = System.currentTimeMillis();
             //SimpleDateFormat sdf = new SimpleDateFormat("MMM dd, yyyy HH:mm");
             //Date currentDate = new Date(currentMillis);
             activeUser.setTimeStamp(currentMillis);
